services/binah-ml/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MLFlow and experiment tracking
mlflow==2.9.2
mlflow[extras]==2.9.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Database
neo4j==5.15.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23

# Utilities
python-multipart==0.0.6
httpx==0.25.2
joblib==1.3.2

# JWT Authentication
python-jose[cryptography]==3.3.0

# Kafka for event consumption
aiokafka==0.8.1
kafka-python==2.0.2
